Move App into Maxiplex
