+++ packages/bigby-core/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./App"
 export * from "./transforms"
-export * from "./types"
