+++ packages/maxiplex-core/src/App.ts
@@ -1,12 +1,15 @@
-import { World } from "@maxiplex/core"
 import { Event } from "@bigby/event"
-import {
-  OnLoadCallback,
-  Plugin,
-  OnStartCallback,
-  OnStopCallback,
-  OnUpdateCallback
-} from "./types"
+import { World } from "./World"
+
+export type OnLoadCallback<A extends App> = (app: A) => void | Promise<void>
+
+export type OnStartCallback<A extends App> = (app: A) => void | Promise<void>
+
+export type OnUpdateCallback = (dt: number) => void
+
+export type OnStopCallback<A extends App> = (app: A) => void
+
+export type Plugin<A extends App> = (app: A) => A
 
 export type BaseEntity = {}
